Added XP Pooling

+++ Organism AI/AbstOrganism.java
@@ -15,14 +15,14 @@ public abstract class AbstOrganism extends Actor {
     //"Live" Variables
     public int age; //Time in seconds
     public int health; //Health
-    public int xp; //XP - allows for RPG-like store for mutation
+    public double xp; //XP - allows for RPG-like store for mutation
     public int radius; //Based off health, to visually see health
     public Boolean isAlive;
     public Boolean isAlpha;
 
     //XP Upgradeable Variables
     public int maxHealth; //Maximum Health
-    public int maxXp; //Max XP Storage
+    public double maxXp; //Max XP Storage
     public int speed; //Nuff Said
     public int att; //Attck power
     public int def; //defensive power
